feat(i18n): modularize translations
