+++ src/i18n/landing.ts
@@ -1,0 +1,13 @@
+export default {
+  "forêt brumeuse": { fr: "forêt brumeuse", en: "misty forest" },
+  "Trouvez vos coins à champignons comestibles, même sans réseau.": {
+    fr: "Trouvez vos coins à champignons comestibles, même sans réseau.",
+    en: "Find your edible mushroom spots, even offline.",
+  },
+  "Voir la carte": { fr: "Voir la carte", en: "See map" },
+  "Les champignons": { fr: "Les champignons", en: "Mushrooms" },
+  "Mini‑pack offline inclus : carte topo (50 km) + fiches Cèpe, Girolle, Morille.": {
+    fr: "Mini‑pack offline inclus : carte topo (50 km) + fiches Cèpe, Girolle, Morille.",
+    en: "Mini offline pack included: topo map (50 km) + sheets Porcini, Chanterelle, Morel.",
+  },
+};
